+++ docs/CHANGELOG.rst
@@ -14,7 +14,8 @@ Added
 -----
 - Add ``resolwebio/slamdunk`` Docker image
 - Add Tabix (1.7-2) to ``resolwebio/bamliquidator:1.2.0`` docker image
-- Add ``seqtk-rev-complement-single``and ``seqtk-rev-complement-paired`` process
+- Add ``seqtk-rev-complement-single`` and ``seqtk-rev-complement-paired`` process
+- Add ``slamdunk-all-paired`` process
 
 Changed
 -------